refactored menu to class Menu, which is just one big comment
